+++ common/src/main/java/com/selfdot/cobblemonmegas/common/item/KeyStoneType.java
@@ -1,0 +1,3 @@
+package com.selfdot.cobblemonmegas.common.item;
+
+public record KeyStoneType(String name, int customModelData) { }