+++ common/src/main/java/com/selfdot/cobblemonmegas/common/util/ItemUtils.java
@@ -1,0 +1,12 @@
+package com.selfdot.cobblemonmegas.common.util;
+
+import net.minecraft.item.ItemStack;
+
+public class ItemUtils {
+
+    public static void setNameNoItalics(ItemStack itemStack, String customName) {
+        itemStack.getOrCreateSubNbt("display")
+            .putString("Name", "{\"text\":\"" + customName + "\",\"italic\":false}");
+    }
+
+}